feat: Improve the visualisation of Employmentstatus field
When we explore the relationship between the target feature
and a categorical variable, it's usefull to put a bar plot
of the categorical variable itself to show the count
difference between the categories, which could make the
comparison easier.

Also, I add a new parameter `facet_ncol` to the `bar_plot`
function to make it more flexible.